update/WF-10 - Update dao layer

+++ src/main/java/com/github/blog/dao/CommentReactionDao.java
@@ -5,5 +5,5 @@ import com.github.blog.model.CommentReaction;
 /**
  * @author Raman Haurylau
  */
-public interface CommentReactionDao extends CrudDao<CommentReaction, Integer> {
+public interface CommentReactionDao extends CrudDao<CommentReaction, Long> {
 }
